feat(frq): Support every FRQ format the source exams use
- Schema (migration 20260924000000_frq_formats, applied): decimal points,
  pointStep (rubric's own granularity), section, orderIndex and
  solutionSource on questions; answerFormat (WORKED / SHORT_ANSWER /
  DRAWING), leadIn and acceptedAnswers on parts; ANSWER_SHEET media and a
  figure key for [[figure:key]] placement; gradingMethod on submissions.
- Grading rounds to the marking scheme's step (whole points unless the
  rubric awards halves), reads one answer box for all parts, judges
  drawing parts from the photographed answer sheet, and keeps comments on
  full-mark parts short while making feedback on lost marks specific.
- Short-answer questions are checked exactly on the server for free
  (checkShortAnswers, POST /api/frq/[id]/check): no AI, no credit, no
  attempt used, and answers are never revealed until every blank is right.
- Parts solved in full show their model solution straight away.
- Student page renders maths with KaTeX, places figures where the text
  refers to them, offers answer-sheet downloads, and labels where the
  solution came from (official, adapted, or written by Astro Coach).
- CLAUDE.md Pro rules updated to match.

Tests: 220/220 (new short-answer and publish-gate tests); tsc and eslint
clean.

+++ src/app/api/admin/frq/drafts/[id]/route.ts
@@ -52,7 +52,7 @@ export async function GET(
 const patchSchema = z.object({
   questionText: z.string().min(1).optional(),
   title: z.string().max(300).nullish(),
-  totalPoints: z.number().int().min(0).optional(),
+  totalPoints: z.number().min(0).optional(),
   officialSolution: z.string().nullish(),
   gradingRubric: z.string().nullish(),
   primaryCurriculumTopic: z.string().min(1).optional(),
@@ -79,7 +79,7 @@ const patchSchema = z.object({
         id: z.string(),
         label: z.string().min(1).max(20),
         prompt: z.string().min(1),
-        maxPoints: z.number().int().min(0),
+        maxPoints: z.number().min(0),
         officialSolution: z.string().nullish(),
         gradingRubric: z.string().nullish(),
       }),
